zsh common- in kill-ssh function, handle condition when given ssh session is not present

+++ ohmyzsh-common/functions.zsh
@@ -59,10 +59,12 @@ function share() {
 
 function kill-ssh() {
     process=$(ps aux | grep ssh | grep "$1" | grep -v grep | head -n1)
-    echo -n "Kill" $process "? (Y/n) "
-    read opt
-    if [[ "$(echo $opt | tr "[A-Z]" "[a-z]")" == "n" ]]; then
-    else
-        sudo kill -INT `echo $process | awk '{print $2}'`
+    if [[ -n "$process" ]]; then
+        echo -n "Kill" $process "? (Y/n) "
+        read opt
+        if [[ "$(echo $opt | tr "[A-Z]" "[a-z]")" == "n" ]]; then
+        else
+            sudo kill -INT `echo $process | awk '{print $2}'`
+        fi
     fi
 }
